ASM-7756 Double quotes in init.pp causing MSSQL install failure
MSSQL installs are failing due to a change in the puppet-mssql2012 init.pp
manifest. On the failing server, the event viewer shows a Puppet ERROR
event complaining about a file path needing to be fully qualified. The file
and path appears in the message text:

'C: ql2012install.ini'

Note that there is a space between 'C:' and 'ql2012install.ini' where an 's'
should be.

This was due to a last minute code change as a result of a review comment.
The reviewer instructed that we should change all single quotes in this file
to double quotes. The problem arises because within double quotes Puppet will
perform interpolation which includes translating escape sequences. In par-
ticular, in this case, the original path included \s (as part of
C:\sql2012install.ini) but \s is translated to a space which is what appears
in the Puppet event viewer error message.